Move is_foolike functions to images.py
The new images.py now houses image-related utility functions.
